requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
requests>=2.31.0
mediapipe>=0.10.0
withoutbg>=1.0.0
